removed get_is_temp_field_also_for_ptr() method and reduced use of set_is_temp_field_one_level()

+++ src/api/staticcontextimpl.h
@@ -52,7 +52,7 @@ class StaticContextImpl : public StaticContext
 protected:
   static_context_t                    theCtx;
 
-  std::map<short, static_context_t>   theSctxMap;
+  std::map<int, static_context_t>     theSctxMap;
 
   DiagnosticHandler                 * theDiagnosticHandler;
   bool                                theUserDiagnosticHandler;
